backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
h3==4.2.2
idna==3.10
ndg-httpsclient==0.5.1
numpy==2.2.4
packaging==24.2
phonenumbers==8.13.49
Pillow==10.1.0
pusher==3.3.3
pyasn1==0.6.1
pycountryinfo==1.0.8
pycparser==2.22
PyJWT==2.10.1
PyNaCl==1.5.0
pyOpenSSL==25.0.0
python-dotenv==1.0.0
pytz==2025.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
timezonefinder==6.5.8
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
whitenoise==6.6.0
